Patch from John Pullan to kill off the dvb signal monitor thread for general useage.
git-svn-id: http://svn.mythtv.org/svn/trunk@5822 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/dvbtypes.h
@@ -76,8 +76,8 @@ typedef struct dvbchannel
 
 typedef struct dvbstats
 {
-    uint16_t snr;
-    uint16_t ss;
+    uint32_t snr;
+    uint32_t ss;
     uint32_t ber;
     uint32_t ub;
